% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADS.R
\name{ADS}
\alias{ADS}
\title{Adaptive Discrete Smoothing Algorithm}
\usage{
ADS(
  df,
  target,
  individ,
  learner = mlr3::mlr_learners$get("regr.lm"),
  delta = 0.7,
  gamma = 1,
  iterations = 2,
  W_start = NULL,
  calc_dist = calc_dist_default,
  kernel = "gaussian",
  parallel = F,
  ...
)
}
\arguments{
\item{df}{A \code{data.frame} containing the covariates,target variable and a factor for the individuum.}

\item{target}{Name of the target column.}

\item{individ}{Name of the individuum column.}

\item{learner}{The specified learner.}

\item{delta}{Weighting of the distance.}

\item{gamma}{Weighting of the distance.}

\item{iterations}{Number of iteration}

\item{W_start}{The specified starting weights.}

\item{calc_dist}{The distance function.}

\item{kernel}{The kernel used in the distance function.}

\item{parallel}{Parallel execution}

\item{...}{Additional arguments.}
}
\value{
An S3 object of class \code{ADS} with components
\itemize{
\item \code{learnerlist} A list of the estimated learners.
\item \code{weight_path} An array containing the different weight matrices.
\item \code{fitted.values} The corresponding predictions.
\item \code{level_ind} A vector containing the indications of the levels (of the individuals.)
\item \code{task_list} A list containing the \code{mlr3} tasks.
\item \code{fit_list} A list of the fits corresponding to the indiviuals.
\item \code{input} A list of important inputs (the name of the column of the individuals.)
}
}
\description{
Adaptive Discrete Smoothing Algorithm
}
\examples{
#create data
n <- 200; N <- 10; p <- 3
X <- matrix(runif(N*n*p),nrow = n*N, ncol = p)
ind <- as.factor(rep(1:N,n)[sample(1:(n*N),n*N)])

#create 2 different types
Y <- X\%*\%rep(1,p) * (as.numeric(ind) <= N/2) + rnorm(n*N,0,1)

data <- data.frame(X,"y" = Y, "ind" = ind)

model <- ADS(df = data,target = "y",individ = "ind")



}
\seealso{
\code{predict.ADS}, \code{autoplot.ADS}
}
