% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADS.R
\name{autoplot.ADS}
\alias{autoplot.ADS}
\title{Plot a heatmap of the weight matrix.}
\usage{
\method{autoplot}{ADS}(object, iterations = 1:2)
}
\arguments{
\item{object}{\code{ADS} object.}

\item{iterations}{Iterations of the weightmatrix to plot (1 equals the input matrix, default is diagonal.)}
}
\value{
A \code{ggplot} object with a heatmap.
}
\description{
Plot a heatmap of the weight matrix.
}
