% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_default.R
\name{calc_dist_default}
\alias{calc_dist_default}
\title{Distance between models.}
\usage{
calc_dist_default(
  model_1,
  model_2,
  gamma = 1,
  task_list,
  kernel = "gaussian",
  ...
)
}
\arguments{
\item{model_1}{First Model.}

\item{model_2}{Second Model.}

\item{gamma}{Weighting of the distance.}

\item{task_list}{Task list.}

\item{kernel}{Kernel}

\item{...}{Additional arguments}
}
\value{
A distance between the two models.
}
\description{
Distance between models.
}
