% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight_default.R
\name{calc_weight_default}
\alias{calc_weight_default}
\title{Calculate weights from the distance.}
\usage{
calc_weight_default(dist, delta = 0.7, gamma = 0.5, kernel = "gaussian", ...)
}
\arguments{
\item{delta}{(\code{numeric(1)}) \cr
Scaling of the distance (bandwidth of the kernel).}

\item{gamma}{(\code{numeric(1)}) \cr
Final scaling of the weight.}

\item{kernel}{(\code{character(1)}) \cr
A kernel of "gaussian", "epa", "unif", "tri".}

\item{...}{Additional arguments}

\item{distance}{(\code{numeric(1)}) \cr
A distance.}
}
\value{
(\code{numeric(1)}) \cr
A weight corresponding to the distance
}
\description{
Calculate weights from the distance.
}
