% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADS.R
\name{predict.ADS}
\alias{predict.ADS}
\title{Predict Method for ADS object.}
\usage{
\method{predict}{ADS}(object, newdata, ...)
}
\arguments{
\item{object}{ADS object.}

\item{newdata}{New dataframe.}

\item{...}{Additional arguments}
}
\value{

}
\description{
Predict Method for ADS object.
}
