% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADS.R
\name{ADS}
\alias{ADS}
\title{ADS class}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Base class for Adaptive discrete smoothing.
}
\concept{ADS}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\code{data.frame()})\cr
Data frame.}

\item{\code{learner}}{(\code{\link[mlr3:LearnerRegr]{LearnerRegr}})\cr
A learner.}

\item{\code{task_list}}{(\code{list()})\cr
A list of tasks.}

\item{\code{learner_list}}{(\code{list()})\cr
A list of learners.}

\item{\code{delta}}{(\code{numeric()})\cr
A vector.}

\item{\code{gamma}}{(\code{numeric()})\cr
A vector.}

\item{\code{weight_path}}{(\code{array()})\cr
An array.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ADS$new()}}
\item \href{#method-print}{\code{ADS$print()}}
\item \href{#method-fit}{\code{ADS$fit()}}
\item \href{#method-predict}{\code{ADS$predict()}}
\item \href{#method-heatmap}{\code{ADS$heatmap()}}
\item \href{#method-clone}{\code{ADS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ADS Class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADS$new(
  data,
  target,
  individ,
  learner,
  delta = 0.7,
  gamma = 1,
  iterations = 2,
  W_start = NULL,
  calc_dist = list(fun = calc_dist_default),
  calc_weight = list(fun = calc_weight_default, params = list(kernel = "gaussian"))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{data.frame()}) \cr
Data.frame containing the target variable, the covariates and a column with the corresponding individuals.}

\item{\code{target}}{(\code{character(1)}) \cr
Name of the target variable.}

\item{\code{individ}}{(\code{character(1)}) \cr
Name of the column with the individuals. Column has to be a factor.}

\item{\code{learner}}{(\code{\link[mlr3:LearnerRegr]{LearnerRegr}}) \cr
The machine learners from the \code{mlr3}-package.}

\item{\code{delta}}{(\code{numeric()}) \cr
Parameter for the loss function}

\item{\code{gamma}}{(\code{gamma()}) \cr
Parameter for the loss function}

\item{\code{iterations}}{(\code{integer(1)}) \cr
Number of iterations.}

\item{\code{W_start}}{(\code{matrix()}) \cr
Starting weights. Default is the identity matrix.}

\item{\code{calc_dist}}{(\verb{function()}) \cr
A list. If not set to default a list with the following components:
\itemize{
\item \code{fun} A function of the form \verb{function(model_1,model_2,task_list,...)} to calculate the weights.
\item \code{params} A list of the additional parameters used in \code{fun}.
}}

\item{\code{calc_weight}}{(\verb{function()}) \cr
A list. If not set to default a list with the following components:
\itemize{
\item \code{fun} A function of the form \verb{function(dist, delta, gamma,...)} to calculate the weights.
\item \code{params} A list of the additional parameters used in \code{fun}.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print ADS objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADS$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Estimate ADS models.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADS$fit(store_predictions = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store_predictions}}{(\code{logical(1)}) \cr
Indicates whether the predictions for the nuisance functions should be
stored in field \code{predictions}. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict ADS models on new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADS$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{(\code{data.frame()}) \cr
Predicts the model on new data. Has to be a data.frame with the same columns as in the trained model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector containig the predicted values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-heatmap}{}}}
\subsection{Method \code{heatmap()}}{
Plot the used weights as a heatmap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADS$heatmap(iterations = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{(\code{integer()}) \cr
Steps to plot the weights for. Defaults to all iterations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ADS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
