% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADS.R
\name{ADS}
\alias{ADS}
\title{Adaptive Discrete Smoothing Algorithm}
\usage{
ADS(
  df,
  target,
  individ,
  learner = mlr3::mlr_learners$get("regr.lm"),
  delta = 0.5,
  gamma = 1,
  iterations = 2,
  W_start = NA,
  calc_dist = calc_dist_default,
  kernel = "gaussian",
  parallel = F,
  ...
)
}
\arguments{
\item{df}{A \code{data.frame} containing the covariates,target variable and a factor for the individuum.}

\item{target}{Name of the target column.}

\item{individ}{Name of the individuum column.}

\item{learner}{The specified learner.}

\item{delta}{Weighting of the distance.}

\item{gamma}{Weighting of the distance.}

\item{iterations}{Number of iteration}

\item{W_start}{The specified starting weights.}

\item{calc_dist}{The distance function.}

\item{kernel}{The kernel used in the distance function.}

\item{parallel}{Parallel execution}

\item{...}{Additional arguments.}
}
\value{

}
\description{
Adaptive Discrete Smoothing Algorithm
}
