% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_default.R
\name{calc_dist_default}
\alias{calc_dist_default}
\title{Distance between models.}
\usage{
calc_dist_default(model_1, model_2, task_list, ...)
}
\arguments{
\item{model_1}{(\code{\link[mlr3:LearnerRegr]{LearnerRegr}})\cr
First Model.}

\item{model_2}{(\code{\link[mlr3:LearnerRegr]{LearnerRegr}})\cr
Second Model.}

\item{task_list}{(\code{\link[mlr3:TaskRegr]{TaskRegr}})\cr
Task list.}

\item{...}{Additional arguments}
}
\value{
(\code{numeric(1)}) \cr
A distance between the two models.
}
\description{
Distance between models.
}
